+++ modules/cpp/common/ExtSdkMethodType.h
@@ -22,6 +22,9 @@ public:
     static const std::string kickAllDevices;
     static const std::string getLoggedInDevicesFromServer;
     static const std::string getCurrentUser;
+    static const std::string bindDeviceToken;
+    static const std::string bindPushKitToken;
+    static const std::string unbindPushKitToken;
 
     static const std::string onConnected;
     static const std::string onDisconnected;