+++ modules/objc/dispatch/ExtSdkDispatch.m
@@ -94,8 +94,17 @@ static NSString *const TAG = @"ExtSdkDispatch";
     case ExtSdkMethodKeyGetCurrentUserValue:
         [[ExtSdkClientWrapper getInstance] getCurrentUser:ps withMethodType:methodType result:callback];
         break;
-    case ExtSdkMethodKeyUpdatePushConfigValue:
-        [[ExtSdkClientWrapper getInstance] updatePushConfig:ps withMethodType:methodType result:callback];
+    case ExtSdkMethodKeyRenewTokenValue:
+        [[ExtSdkClientWrapper getInstance] renewToken:ps withMethodType:methodType result:callback];
+        break;
+    case ExtSdkMethodKeyBindDeviceTokenValue:
+        [[ExtSdkClientWrapper getInstance] bindDeviceToken:ps withMethodType:methodType result:callback];
+        break;
+    case ExtSdkMethodKeyBindPushKitTokenValue:
+        [[ExtSdkClientWrapper getInstance] bindPushKitToken:ps withMethodType:methodType result:callback];
+        break;
+    case ExtSdkMethodKeyUnbindPushKitTokenValue:
+        [[ExtSdkClientWrapper getInstance] unbindPushKitToken:ps withMethodType:methodType result:callback];
         break;
 
     /// #pragma mark - EMClientDelegate value